server/pyproject.toml
[project]
name = "nancysgun"
version = "1.0.0"
description = "POS framework"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "argon2-cffi>=25.1.0",
    "fastapi>=0.135.3",
    "openpyxl>=3.1.5",
    "pandas>=3.0.2",
    "passlib[argon2]>=1.7.4",
    "pydantic-settings>=2.13.1",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.24",
    "sqlalchemy>=2.0.49",
    "uvicorn>=0.44.0",
    "zxcvbn>=4.5.0",
]
